Rename useFacets file
